refactor(functions): organize handlers into modules
